% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{alotau_accord_2014}
\alias{alotau_accord_2014}
\title{alotau_accord_2014}
\format{A dataframe with 477 rows and 2 columns:
\describe{
\item{\code{linenumber}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\source{
\url{https://srhr.org/abortion-policies/documents/countries/05-PNG-Policy-on-Free-Primary-Health-Care-an-Subsidized-Specialized-Health-Services-in-PNG-Department-of-Health-2013.pdf}
}
\usage{
alotau_accord_2014
}
\description{
Papua New Guinea's Alotau Accord for a Government of Reconciliation, Unity and Hope
}
\details{
A dataset containing the text of Papua New Guinea's Alotau Accord for a
Government of Reconciliation, Unity and Hope. The UTF-8 plain text is
divided into elements of up to about 70 characters each. Some elements are
blank.
}
\keyword{datasets}
