% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{country}
\alias{country}
\title{country}
\format{A SpatialPolygonsDataFrame with 1 feature with 2 variables:
\describe{
\item{\code{ADM0_EN}}{Country Name; Papua New Guinea}
\item{\code{ADM0_PCODE}}{Two-character country code; PG}
}}
\source{
\href{Humanitarian Data Exchange}{https://data.humdata.org/dataset/papua-new-guinea-admin-level-0-international-boundaries}
}
\usage{
country
}
\description{
Papua New Guinea Country Borders
}
\keyword{datasets}
