% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{district}
\alias{district}
\title{district}
\format{A SpatialPolygonsDataFrame with 87 features each with 6 variables:
\describe{
\item{\code{ADM2_EN}}{District Name}
\item{\code{ADM2_PCODE}}{Four character district code}
\item{\code{ADM1_EN}}{Province Name}
\item{\code{ADM1_PCODE}}{Two character province code}
\item{\code{ADM0_EN}}{Country Name; Papura New Guinea}
\item{\code{ADM0_PCODE}}{Two-character country code; PG}
}}
\source{
\href{Humanitarian Data Exchange}{https://data.humdata.org/dataset/papua-new-guinea-admin-level-0-international-boundaries}
}
\usage{
district
}
\description{
Papua New Guinea District Borders
}
\keyword{datasets}
