% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{health_sector_gender_policy_2014}
\alias{health_sector_gender_policy_2014}
\title{health_sector_gender_policy_2014}
\format{A character vector with 698 elements}
\usage{
health_sector_gender_policy_2014
}
\description{
Papua New Guinea's Health Sector Gender Policy 2014
}
\details{
A dataset containing the text of Papua New Guinea's Health Sector Gender
Policy 2014. The UTF-8 plain text is divided into elements of up to about 70
characters each. Some elements are blank.
}
\keyword{datasets}
