% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{healthfacility}
\alias{healthfacility}
\title{healthfacility}
\format{A data frame with 759 rows and 12 columns:
\describe{
\item{\code{NAME}}{Name of health facility}
\item{\code{TYPE}}{Type of health facility}
\item{\code{LLG_Pcode}}{Local level government code}
\item{\code{LLG_Name}}{Local level government name}
\item{\code{Dis_Pcode}}{District code}
\item{\code{District_Name}}{District name}
\item{\code{Prov_Pcode}}{Province code}
\item{\code{Prov_name}}{Province name}
\item{\code{X}}{Longitude coordinate of health facility}
\item{\code{Y}}{Latitude coordinate of health facility}
\item{\code{Source}}{Source of data}
\item{\code{Feature}}{Features}
}}
\source{
\href{Humanitarian Data Exchange}{https://data.humdata.org/dataset/papua-new-guinea-health-facilities}
}
\usage{
healthfacility
}
\description{
Geolocations of health facilities in Papua New Guinea
}
\keyword{datasets}
