% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{llg}
\alias{llg}
\title{llg}
\format{A SpatialPolygonsDataFrame with 326 features each with 8 variables:
\describe{
\item{\code{ADM3_EN}}{Local level government Name}
\item{\code{ADM3_PCODE}}{Six character local level government code}
\item{\code{ADM2_EN}}{District Name}
\item{\code{ADM2_PCODE}}{Four character district code}
\item{\code{ADM1_EN}}{Province Name}
\item{\code{ADM1_PCODE}}{Two character province code}
\item{\code{ADM0_EN}}{Country Name; Papura New Guinea}
\item{\code{ADM0_PCODE}}{Two-character country code; PG}
}}
\source{
Humanitarian Data Exchange (https://data.humdata.org/dataset/papua-new-guinea-admin-level-0-international-boundaries)
}
\usage{
llg
}
\description{
Papua New Guinea Local Level Government Borders
}
\keyword{datasets}
