% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{package}
\name{papuanewguinea}
\alias{papuanewguinea}
\alias{papuanewguinea-package}
\title{papuanewguinea}
\description{
OMNeoHealth Datasets for Papua New Guinea
}
\details{
This package contains various datasets collected for the Oxford Maternal and
Neonatal Health project in Papua New Guinea.
}
