% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{png_vision_2015}
\alias{png_vision_2015}
\title{png_vision_2015}
\format{A dataframe with 2819 rows and 2 columns:
\describe{
\item{\code{line}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\usage{
png_vision_2015
}
\description{
Papua New Guinea Vision 2050
}
\details{
A dataset containing the text of Papua New Guinea Vision 2050. The UTF-8
plain text is divided into elements of up to about 70 characters each. Some
elements are blank.
}
\keyword{datasets}
