% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{pop_adm3}
\alias{pop_adm3}
\title{pop_adm3}
\format{A data frame with 326 rows and 15 columns:
\describe{
\item{\code{ADM3_EN}}{Local level government name}
\item{\code{ADM3_PCODE}}{Local level government code}
\item{\code{ADM2_EN}}{District name}
\item{\code{ADM2_PCODE}}{District code}
\item{\code{ADM1_EN}}{Province name}
\item{\code{ADM1_PCODE}}{Province code}
\item{\code{ADM0_EN}}{Country name; Papua New Guinea}
\item{\code{ADM0_PCODE}}{Country code; PG}
\item{\code{WARDS}}{Number of wards}
\item{\code{UNITS}}{Number of units}
\item{\code{HOUSEHOLDS}}{Total number of households}
\item{\code{PERSONS}}{Total number of persons}
\item{\code{MALES}}{Total number of males}
\item{\code{FEMALES}}{Total number of females}
\item{\code{WRA}}{Total women of reproductive age (estimated at 52.4\% of total number of females)}
}}
\source{
\href{Humanitarian Data Exchange}{https://data.humdata.org/dataset/papua-new-guinea-administrative-level-0-1-2-and-3-population-statistics}
}
\usage{
pop_adm3
}
\description{
Population data for local level governments of Papua New Guinea
}
\keyword{datasets}
