% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{province}
\alias{province}
\title{province}
\format{A SpatialPolygonsDataFrame with 22 features each with 4 variables:
\describe{
\item{\code{ADM1_EN}}{Province Name}
\item{\code{ADM1_PCODE}}{Two character province code}
\item{\code{ADM0_EN}}{Country Name; Papura New Guinea}
\item{\code{ADM0_PCODE}}{Two-character country code; PG}
}}
\source{
\href{Humanitarian Data Exchange}{https://data.humdata.org/dataset/papua-new-guinea-admin-level-0-international-boundaries}
}
\usage{
province
}
\description{
Papua New Guinea Province Borders
}
\keyword{datasets}
