% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{taskforce_maternal_health_2009}
\alias{taskforce_maternal_health_2009}
\title{taskforce_maternal_health_2009}
\format{A dataframe with 3178 rows and 2 columns:
\describe{
\item{\code{line}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\usage{
taskforce_maternal_health_2009
}
\description{
Ministerial Taskforce on Maternal Health in Papua New Guinea
}
\details{
A dataset containing the text of Ministerial Taskforce on Maternal Health in
Papua New Guinea. The UTF-8 plain text is divided into elements of up to about
70 characters each. Some elements are blank.
}
\keyword{datasets}
