% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{youth_adolescents_health_policy_2014}
\alias{youth_adolescents_health_policy_2014}
\title{youth_adolescents_health_policy_2014.rda}
\format{A dataframe with 2 columns and 1128 rows:
\describe{
\item{\code{linenumber}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\usage{
youth_adolescents_health_policy_2014
}
\description{
National Youth and Adolescents Health Policy 2014
}
\details{
A dataset containing the text of National Youth and Adolescents Health Policy
2014. The UTF-8 plain text is divided into elements of up to about
70 characters each. Some elements are blank.
}
\keyword{datasets}
