% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{national_sexual_reproductive_policy}
\alias{national_sexual_reproductive_policy}
\title{national_sexual_reproductive_policy}
\format{A dataframe with 712 rows and 2 elements
\describe{
\item{\code{line}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\usage{
national_sexual_reproductive_policy
}
\description{
Papua New Guinea's National Sexual Reproductive Policy
}
\details{
A dataset containing the text of Papua New Guinea's National Sexual Reproductive
Policy. The UTF-8 plain text is divided into elements of up to about 70
characters each. Some elements are blank.
}
\keyword{datasets}
