% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\docType{data}
\name{newborn_health_policy_2014}
\alias{newborn_health_policy_2014}
\title{newborn_health_policy_2014}
\format{A dataframe with 1331 rows and 2 columns
\describe{
\item{\code{line}}{Line number}
\item{\code{text}}{Character string of about 70 characters each}
}}
\usage{
newborn_health_policy_2014
}
\description{
Papua New Guinea's Newborn Health Policy 2014
}
\details{
A dataset containing the text of Papua New Guinea's Newborn Health Policy
2014. The UTF-8 plain text is divided into elements of up to about 70
characters each. Some elements are blank.
}
\keyword{datasets}
